* Changes the time when the lockdown occurs / the spawnpoints are enabled to after the transfer shuttle has departed.

+++ code/modules/ghostroles/spawner/human/admin.dm
@@ -307,30 +307,33 @@
 	qdel(M)
 
 
-/hook/crew_transfer/proc/spawn_odin_staff()
-	if(get_security_level() == "green" && prob(70))
-		var/datum/ghostspawner/G = SSghostroles.get_spawner("odinchef")
-		G.enable()
-		G = SSghostroles.get_spawner("odinbartender")
-		G.enable()
 
-	if(get_security_level() == "blue")
-		var/datum/wifi/sender/door/wifi_sender = new("odin_arrivals_lockdown", SSghostroles)
-		wifi_sender.activate("close")
-		qdel(wifi_sender)
+/hook/shuttle_moved/proc/setup_odin_for_shuttle(var/obj/effect/shuttle_landmark/start_location, var/obj/effect/shuttle_landmark/destination)
+	if(start_location.landmark_tag != "nav_emergency_start")
+		return TRUE
 
-		if(prob(50))
-			var/datum/ghostspawner/G = SSghostroles.get_spawner("checkpointsec")
+	if(emergency_shuttle.evac)
+		if(get_security_level() != "green" )
+			var/datum/wifi/sender/door/wifi_sender = new("odin_arrivals_lockdown", SSghostroles)
+			wifi_sender.activate("close")
+			qdel(wifi_sender)
+
+			if(prob(80))
+				var/datum/ghostspawner/G = SSghostroles.get_spawner("checkpointsec")
+				G.enable()
+		return TRUE
+	else
+		if(get_security_level() == "green" && prob(70))
+			var/datum/ghostspawner/G = SSghostroles.get_spawner("odinchef")
 			G.enable()
-	return TRUE
-
-/hook/emergency_evac/proc/enable_checkpoint()
-	if(get_security_level() != "green" )
-		var/datum/wifi/sender/door/wifi_sender = new("odin_arrivals_lockdown", SSghostroles)
-		wifi_sender.activate("close")
-		qdel(wifi_sender)
-
-		if(prob(80))
-			var/datum/ghostspawner/G = SSghostroles.get_spawner("checkpointsec")
+			G = SSghostroles.get_spawner("odinbartender")
 			G.enable()
-	return TRUE
+		if(get_security_level() == "blue")
+			var/datum/wifi/sender/door/wifi_sender = new("odin_arrivals_lockdown", SSghostroles)
+			wifi_sender.activate("close")
+			qdel(wifi_sender)
+
+			if(prob(50))
+				var/datum/ghostspawner/G = SSghostroles.get_spawner("checkpointsec")
+				G.enable()
+		return TRUE
